requirements.txt
sentence-transformers
faiss-cpu
pandas
numpy
scikit-learn
gradio
folium
openai
transformers
sentencepiece
accelerate

